feat: add logger and refactor big index.ts in smaller modules. closes #28
