+++ index.ts
@@ -1,1098 +1,5 @@
-import Meerkat from '@fabianbormann/meerkat';
-import type {
-  PeerConnectApi,
-  DAppPeerConnectParameters,
-  Cip30Api,
-  Cip30Function,
-  Cbor,
-  Paginate,
-  Bytes,
-  Cip30DataSignature,
-  IConnectMessage,
-  IDAppInfos,
-  IWalletInfo,
-} from './types';
-import QRCode from 'qrcode-svg';
-import { identicon } from '@basementuniverse/marble-identicons';
-import {
-  Value,
-  ExperimentalContainer,
-  buildApiCalls,
-  createTypeMapping,
-  serializeTypeMapping,
-  registerExperimentalEndpoint,
-} from './lib/ExperimentalContainer';
-import { LogLevel } from '@fabianbormann/meerkat/dist/types';
-
-class DAppPeerConnect {
-  private meerkat: Meerkat;
-  private walletDiscoveryMeerkat: Meerkat | null = null
-
-  private connectedWallet: string | null = null;
-  protected enableLogging: boolean = false;
-  protected logger: InfoLogger;
-  protected logLevel: string = 'info'
-
-  protected readonly dAppInfo: IDAppInfos
-
-  protected identicon: string | null = null;
-
-  protected onConnect?: (address: string) => void;
-  protected onDisconnect?: (address: string) => void;
-  protected onApiEject?: (name: string, address: string) => void;
-  protected onApiInject?: (name: string, address: string) => void;
-
-  protected setUpDiscoveryMeerkcat = (address?: string) => {
-
-    if(address || AutoConnectHelper.getWalletDiscoveryAddress()) {
-
-      this.meerkat.logger.debug('DApp: create discovery with address', address ?? AutoConnectHelper.getWalletDiscoveryAddress())
-      this.meerkat.logger.debug('DApp: create discovery with seed',    AutoConnectHelper.getWalletAutoDiscoverySeed())
-
-      this.walletDiscoveryMeerkat = new Meerkat({
-        seed: AutoConnectHelper.getWalletAutoDiscoverySeed() ?? undefined,
-        announce: [
-          'https://pro.passwordchaos.gimbalabs.io',
-          'wss://tracker.files.fm:7073/announce',
-          'wss://tracker.btorrent.xyz',
-          'wss://tracker.openwebtorrent.com:443/announce',
-        ],
-        loggingEnabled: this.enableLogging,
-        identifier: address ?? AutoConnectHelper.getWalletDiscoveryAddress()!
-      }).setMaxListeners(20)
-      this.walletDiscoveryMeerkat.logger.logLevel = this.logLevel as LogLevel
-
-      this.meerkat.logger.debug('DApp: walletDiscoveryMeerkat address:', this.walletDiscoveryMeerkat.address())
-
-      AutoConnectHelper.saveWalletAutoDiscoverySeed(this.walletDiscoveryMeerkat.seed)
-
-      this.meerkat.logger.debug("DApp: Adding onServer event for discover wallet discovery meerkat.")
-
-      this.walletDiscoveryMeerkat.on('server', () => {
-
-        this.meerkat.logger.debug('DApp: SERVER discovery: received on server event')
-
-        if(!this.walletDiscoveryMeerkat) {
-
-          throw new Error('Meerkat not connected.')
-        }
-
-        this.meerkat.logger.debug('DApp: SERVER discovery: Calling rpc connect on wallet.')
-
-        this.walletDiscoveryMeerkat.rpc(AutoConnectHelper.getWalletDiscoveryAddress()!, 'connect', { dappAddress: this.meerkat.address() }, (connectStatus: any) => {
-
-          this.meerkat.logger.debug('DApp: SERVER discovery: Client connect status: ', connectStatus)
-
-        });
-      })
-
-    }
-  }
-
-  public setLogLevel = (level: string, meerkat: boolean = false) => {
-
-    this.logLevel = level
-
-    this.logger.setLevel(level)
-
-    if(this.meerkat && meerkat) {
-      this.meerkat.logger.logLevel = level as LogLevel
-    }
-
-    if(this.walletDiscoveryMeerkat && meerkat) {
-      this.walletDiscoveryMeerkat.logger.logLevel = level as LogLevel
-    }
-  }
-
-  constructor({
-    dAppInfo,
-    seed,
-    announce,
-    loggingEnabled,
-    verifyConnection,
-    onConnect,
-    onDisconnect,
-    onApiEject,
-    onApiInject,
-    useWalletDiscovery
-  }: DAppPeerConnectParameters) {
-    this.dAppInfo = dAppInfo;
-
-    if(loggingEnabled) {
-      this.enableLogging = loggingEnabled
-    }
-
-    this.meerkat = new Meerkat({
-      seed: seed || localStorage.getItem('meerkat-dapp-seed') || undefined,
-      announce: announce,
-      loggingEnabled: loggingEnabled,
-    }).setMaxListeners(20);
-
-    this.logger = new InfoLogger('DAppPeerConnect',loggingEnabled, 'info' );
-    this.meerkat.logger.logLevel = this.logLevel as LogLevel
-
-    if(useWalletDiscovery) {
-
-      setTimeout(() => {
-
-        //initialize discovery meerkat 1 second later
-        this.setUpDiscoveryMeerkcat()
-
-      }, 1000)
-
-    }
-
-    this.onConnect = onConnect;
-    this.onDisconnect = onDisconnect;
-    this.onApiEject = onApiEject;
-    this.onApiInject = onApiInject;
-
-    localStorage.setItem('meerkat-dapp-seed', this.meerkat.seed);
-
-    this.logger.info(
-      `The generated meerkat address is: ${this.meerkat.address()}`
-    );
-
-    this.dAppInfo.address = this.meerkat.address();
-
-    let connected = false;
-
-    this.meerkat.on('connections', () => {
-      if (!connected) {
-        connected = true;
-        this.logger.info('server ready');
-      }
-    });
-
-    this.meerkat.on('seen', (address) => {
-      const globalCardano = (window as any).cardano || {};
-      if (
-        Object.keys(globalCardano).find(
-          (apiName) => globalCardano[apiName].identifier === address
-        )
-      ) {
-        this.logger.info(`Saw address ${address}`);
-      } else {
-        this.logger.info(
-          `Saw address ${address} but it has not injected it's api yet`
-        );
-      }
-    });
-
-    this.meerkat.on('left', (address: string) => {
-      this.leftServer(address);
-    });
-
-    this.meerkat.register(
-      'connect',
-      (
-        address: string,
-        walletInfo: IWalletInfo,
-        callback: (args: IConnectMessage) => void
-      ) => {
-        if (!this.connectedWallet) {
-          const connectWallet = (
-            granted: boolean,
-            allowAutoConnect: boolean = false
-          ) => {
-            if (walletInfo.requestAutoconnect && granted && allowAutoConnect) {
-              AutoConnectHelper.addAutoConnectId(address);
-            }
-
-            if (granted) {
-              this.connectedWallet = address;
-              this.logger.info(
-                `Successfully connected ${this.connectedWallet}`
-              );
-
-              callback({
-                dApp: this.dAppInfo,
-                address: address,
-                connected: true,
-                error: false,
-                autoConnect: allowAutoConnect,
-              });
-
-              this.generateIdenticon();
-
-              if (this.onConnect) {
-                this.onConnect(address);
-              }
-            } else {
-              callback({
-                dApp: this.dAppInfo,
-                address: address,
-                connected: false,
-                error: true,
-                errorMessage: `User denied connection to ${address}`,
-                autoConnect: allowAutoConnect,
-              });
-
-              this.logger.info(`User denied connection to ${address}`);
-            }
-          };
-
-          if (typeof verifyConnection !== 'undefined') {
-            if (AutoConnectHelper.isAutoConnectId(address)) {
-              connectWallet(true);
-            } else {
-              verifyConnection(
-                {
-                  ...walletInfo,
-                  address: address,
-                },
-                connectWallet
-              );
-            }
-          } else {
-            connectWallet(true);
-          }
-        } else if (this.connectedWallet === address) {
-          this.logger.info(
-            `Connection has already been established to ${address}.`
-          );
-
-          callback({
-            address: address,
-            dApp: this.dAppInfo,
-            connected: true,
-            error: false,
-          });
-        } else {
-          callback({
-            dApp: this.dAppInfo,
-            address: address,
-            connected: false,
-            error: false,
-            errorMessage:
-              'Connection failed. Another wallet has already been connected to this dApp.',
-          });
-
-          this.logger.info(
-            'Connection failed. Another wallet has already been connected to this dApp.'
-          );
-        }
-      }
-    );
-
-    /**
-     * Client signals that it is disconnecting
-     */
-    this.meerkat.register(
-      'disconnect',
-
-      (
-        address: string,
-        walletInfo: IWalletInfo,
-        callback: (args: IConnectMessage) => void
-      ) => {
-        if (this.connectedWallet) {
-          if (this.connectedWallet !== address) {
-            this.logger.info(
-              `Unregistered address ${address} is calling disconnect.`
-            );
-
-            callback({
-              dApp: this.dAppInfo,
-              connected: false,
-              error: true,
-              errorMessage:
-                'Unregistered address ${address} is calling disconnect.',
-            });
-
-            return;
-          }
-
-          this.logger.info(
-            `Wallet ${this.connectedWallet} is calling disconnect.`
-          );
-
-          callback({
-            dApp: this.dAppInfo,
-            connected: false,
-            error: false,
-          });
-
-          this.leftServer(address);
-          this.connectedWallet = null;
-
-          return;
-        }
-
-        this.logger.info(`Calling disconnect with no connected wallet.`);
-
-        callback({
-          dApp: this.dAppInfo,
-          connected: false,
-          error: true,
-          errorMessage: 'No wallet is connected.',
-        });
-      }
-    );
-
-    this.meerkat.register(
-      'setDiscovery',
-      (address: string, args: { walletDiscoveryAddress: string }, callback: (args: boolean) => void) => {
-
-        this.logger.debug('DApp: SERVER: setDiscovery with:', args)
-
-        if(useWalletDiscovery) {
-
-          AutoConnectHelper.saveWalletDiscoveryAddress(args.walletDiscoveryAddress)
-
-          return callback(true)
-        } else {
-
-          return callback(false)
-        }
-      })
-
-    this.meerkat.register(
-      'api',
-      (
-        address: string,
-        args: { api: PeerConnectApi },
-        callback: (args: IConnectMessage) => void
-      ) => {
-        if (address !== this.connectedWallet) {
-          return;
-        }
-
-        const injectedClients = this.getInjectedApis();
-        if (injectedClients.includes(address)) {
-          this.logger.info(`${address} already injected`);
-          return;
-        }
-
-        const api: {
-          [key in Cip30Function | 'experimental']?:
-            | Function
-            | Record<string, Value>;
-        } = {};
-
-        for (const method of args.api.methods) {
-          api[method] = (...params: Array<any>) => {
-            return new Promise((resolve, reject) => {
-              if (typeof params === 'undefined') {
-                params = [];
-              }
-
-              this.meerkat.rpc(
-                address,
-                'invoke',
-                [method, ...params],
-                (result: any) => resolve(result)
-              );
-            });
-          };
-        }
-
-        const initialExperimentalApi = buildApiCalls(
-          this.meerkat,
-          address,
-          args.api.experimentalApi,
-          'invokeExperimental'
-        );
-        const fullExperimentalApi = buildApiCalls(
-          this.meerkat,
-          address,
-          args.api.fullExperimentalApi,
-          'invokeEnableExperimental'
-        );
-
-        api['experimental'] = fullExperimentalApi;
-
-        const cip30Api: Cip30Api = {
-          apiVersion: args.api.apiVersion,
-          name: args.api.name,
-          icon: args.api.icon,
-          identifier: address,
-          experimental: initialExperimentalApi,
-          isEnabled: () => new Promise((resolve, reject) => resolve(true)),
-          enable: () => new Promise((resovle, reject) => resovle(api)),
-        };
-
-        if (this.isWalletNameInjected(args.api.name)) {
-          this.logger.info(
-            `Not injecting wallet api. API for wallet '${args.api.name}' is already injected.`
-          );
-          return callback({
-            dApp: this.dAppInfo,
-            connected: false,
-            error: true,
-            errorMessage: `Wallet with name ${args.api.name} is already injected.`,
-          });
-        }
-
-        if (!this.isP2pWalletCompliantName(args.api.name)) {
-          this.logger.warn(
-            `Injected wallet does not contain 'p2p' in name, this is discouraged. `
-          );
-        }
-
-        (window as any).cardano = (window as any).cardano || {};
-        (window as any).cardano[args.api.name.toLowerCase()] = cip30Api;
-        this.logger.info(
-          `injected api of ${args.api.name} into window.cardano`
-        );
-
-        if (onApiInject) {
-          onApiInject(args.api.name, address);
-        }
-      }
-    );
-  }
-
-  private leftServer = (address: string) => {
-    if (address === this.connectedWallet) {
-      this.connectedWallet = null;
-
-      if (this.onDisconnect) {
-        this.onDisconnect(address);
-      }
-
-      const globalCardano = (window as any).cardano || {};
-      const apiName = Object.keys(globalCardano).find(
-        (apiName) => globalCardano[apiName].identifier === address
-      );
-      if (apiName) {
-        this.logger.info(
-          `${this.connectedWallet} disconnected. ${apiName} has been removed from the global window object`
-        );
-        delete (window as any).cardano[apiName.toLowerCase()];
-        if (this.onApiEject) {
-          this.onApiEject(apiName, address);
-        }
-      } else {
-        this.logger.info(
-          `${this.connectedWallet} disconnected. Cleanup was not necessary.`
-        );
-      }
-    }
-  };
-
-  public shutdownServer = () => {
-    if (this.connectedWallet) {
-      const status: IConnectMessage = {
-        connected: false,
-        error: false,
-        errorMessage: 'Server is closing connections.',
-        dApp: this.dAppInfo,
-      };
-
-      this.meerkat.rpc(this.connectedWallet, 'shutdown', status, () => {});
-    }
-  };
-
-  private getInjectedApis() {
-    const globalCardano = (window as any).cardano || {};
-    return Object.keys(globalCardano)
-      .filter((client) => typeof globalCardano[client].identifier === 'string')
-      .map((client) => globalCardano[client].identifier);
-  }
-
-  /**
-   * Checks if wallet with name is already injected into global cardano namespace.
-   * @param name
-   */
-  private isWalletNameInjected = (name: string) => {
-    const globalCardano = (window as any).cardano || {};
-
-    return Object.keys(globalCardano).find(
-      (apiName) => apiName === name.toLowerCase()
-    );
-  };
-
-  /**
-   * Checks if wallet name contains the string p2p to distinguish from other injection.
-   * @param name
-   */
-  private isP2pWalletCompliantName = (name: string) => {
-    return name.includes('p2p');
-  };
-
-  generateQRCode(canvas: HTMLElement) {
-    const data = `${this.meerkat.address()}:meerkat:${new Date().getTime()}`;
-    var qrcode = new QRCode({
-      content: data,
-      padding: 4,
-      width: 256,
-      height: 256,
-      color: '#000000',
-      background: '#ffffff',
-      ecl: 'M',
-    });
-    canvas.innerHTML = qrcode.svg();
-  }
-
-  getConnectedWallet() {
-    return this.connectedWallet;
-  }
-
-  getAddress() {
-    return this.meerkat.address();
-  }
-
-  getSeed() {
-    return this.meerkat.seed;
-  }
-
-  public generateIdenticon = () => {
-    this.identicon = PeerConnectIdenticon.getBase64Identicon(
-      this.connectedWallet + this.getAddress()
-    );
-  };
-
-  public getIdenticon = () => {
-    return this.identicon;
-  };
-}
-
-abstract class CardanoPeerConnect {
-  protected meerkats: Array<Meerkat> = [];
-  protected walletInfo: IWalletInfo
-  protected onConnect:                  (connectMessage: IConnectMessage) => void;
-  protected onDisconnect:               (connectMessage: IConnectMessage) => void;
-  protected onServerShutdown:           (connectMessage: IConnectMessage) => void;
-  protected onApiInject:                (connectMessage: IConnectMessage) => void;
-  protected identicon: string | null = null;
-  protected meerkat : Meerkat | null = null;
-  protected logLevel: LogLevel = 'info';
-
-  protected DAppDiscoveryMeerkat : Meerkat | null = null;
-  protected _cip30ExperimentalApi?: ExperimentalContainer<any>;
-  protected _cip30EnableExperimentalApi?: ExperimentalContainer<any>;
-
-  protected seed : string | null;
-  protected discoverySeed : string | null;
-  protected announceEndpoints: string[];
-
-  constructor(walletInfo: IWalletInfo, args: {
-    seed?: string | null,
-    announce?: string[],
-    discoverySeed?: string | null,
-    logLevel?: LogLevel
-  } = {}) {
-
-    this.walletInfo           = walletInfo;
-
-    this.seed                 = args.seed ?? null;
-    this.discoverySeed        = args.discoverySeed ?? null;
-    this.announceEndpoints    = args.announce ?? [
-      'https://pro.passwordchaos.gimbalabs.io',
-      'wss://tracker.files.fm:7073/announce',
-      'wss://tracker.btorrent.xyz',
-      'ws://tracker.files.fm:7072/announce',
-      'wss://tracker.openwebtorrent.com:443/announce',
-    ];
-    this.logLevel             = args.logLevel ?? 'info';
-
-    this.onConnect            = (connectMessage: IConnectMessage) => {};
-    this.onDisconnect         = (connectMessage: IConnectMessage) => {};
-    this.onServerShutdown     = () => {};
-    this.onApiInject          = () => {};
-
-    this.setUpDiscoveryMeerkat()
-  }
-
-  protected setLogLevel = (level: LogLevel) => {
-
-    this.logLevel = level
-
-    if(this.meerkat) {
-      this.meerkat.logger.logLevel = level
-    }
-    if(this.DAppDiscoveryMeerkat) {
-      this.DAppDiscoveryMeerkat.logger.logLevel = level
-    }
-  }
-
-  protected setUpDiscoveryMeerkat = () => {
-
-    this.DAppDiscoveryMeerkat = new Meerkat({
-      announce: this.announceEndpoints,
-      seed: this.discoverySeed ?? AutoConnectHelper.getWalletAutoDiscoverySeed() ?? undefined,
-      loggingEnabled: true,
-    }).setMaxListeners(20);
-
-    this.clearSeen()
-
-    this.DAppDiscoveryMeerkat.logger.logLevel = this.logLevel
-
-    this.DAppDiscoveryMeerkat?.logger.debug('WALLET: discovery address:', this.DAppDiscoveryMeerkat.address())
-
-    if(!this.discoverySeed) {
-
-      AutoConnectHelper.saveWalletAutoDiscoverySeed(this.DAppDiscoveryMeerkat.seed)
-    }
-
-    this.DAppDiscoveryMeerkat.register(
-      'connect',
-      (address: string, params: { dappAddress: string } , callback: (args: any) => void) => {
-
-        this.DAppDiscoveryMeerkat?.logger.debug('Wallet: DApp is connecting to discovery server!', params.dappAddress)
-
-        this.connect(params.dappAddress)
-
-        callback(true)
-      }
-    )
-
-    this.meerkats.push(this.DAppDiscoveryMeerkat);
-  }
-
-  public setOnConnect = (
-    onConnectCallback: (connectMessage: IConnectMessage) => void
-  ) => {
-    this.onConnect = onConnectCallback;
-  };
-
-  public setOnDisconnect = (
-    onDisconnectCallback: (connectMessage: IConnectMessage) => void
-  ) => {
-    this.onDisconnect = onDisconnectCallback;
-  };
-
-  public setOnServerShutdown = (
-    onServerShutdown: (connectMessage: IConnectMessage) => void
-  ) => {
-    this.onServerShutdown = onServerShutdown;
-  };
-
-  public setOnApiInject = (
-    onApiInject: (connectMessage: IConnectMessage) => void
-  ) => {
-    this.onApiInject = onApiInject;
-  };
-
-  public setExperimentalApi<T extends Record<keyof T, Value>>(
-    dynamicObj: ExperimentalContainer<T>
-  ): void {
-    this._cip30ExperimentalApi = dynamicObj;
-  }
-
-  public setEnableExperimentalApi<T extends Record<keyof T, Value>>(
-    dynamicObj: ExperimentalContainer<T>
-  ): void {
-    this._cip30EnableExperimentalApi = dynamicObj;
-  }
-
-  public getMeercat(identifier: string): Meerkat | undefined {
-    return this.meerkats.find((meerkat) => meerkat.identifier === identifier);
-  }
-
-  public clearSeen = () => {
-
-    if(this.meerkat) {
-      this.meerkat.logger.debug('WALLET: meerkat clear all seen clients.')
-      this.meerkat.seen = {}
-    }
-
-    if(this.DAppDiscoveryMeerkat) {
-      this.DAppDiscoveryMeerkat.logger.debug('WALLET: discovery meerkat clear all seen clients.')
-      this.DAppDiscoveryMeerkat.seen = {}
-    }
-
-  }
-
-  public connect(identifier: string): string {
-
-    this.meerkat = new Meerkat({
-      identifier: identifier,
-      announce: this.announceEndpoints,
-      seed: this.seed ?? undefined,
-    }).setMaxListeners(20);
-    this.meerkat.logger.logLevel = this.logLevel
-
-    this.meerkat?.logger.debug('WALLET: calling to connect to DApp:', identifier)
-
-
-    this.meerkat.register(
-      'shutdown',
-      async (address: string, args: IConnectMessage, callback: Function) => {
-        if (address !== args.dApp.address) {
-          throw new Error(
-            `Address ${args.address} tries to send shutdown for server, ${args.address}.`
-          );
-        }
-
-        this.clearSeen();
-        this.onServerShutdown(args);
-
-        this.meerkat?.logger.debug('WALLET: setup discovery again after server did shutdown.')
-
-        this.setUpDiscoveryMeerkat();
-      })
-
-    this.meerkat.register(
-      'invoke',
-      async (address: string, args: Array<any>, callback: Function) => {
-        const cip30Function = args[0] as Cip30Function;
-
-        if (address === identifier) {
-          const result = await (<any>this[cip30Function])(...args.splice(1));
-          if (typeof result !== 'undefined') {
-            callback(result);
-          }
-        }
-      }
-    );
-
-    registerExperimentalEndpoint(
-      this.meerkat,
-      'invokeExperimental',
-      this._cip30ExperimentalApi!,
-      identifier
-    );
-    registerExperimentalEndpoint(
-      this.meerkat,
-      'invokeEnableExperimental',
-      this._cip30EnableExperimentalApi!,
-      identifier
-    );
-
-    const injectApi = () => {
-      if (!this.meerkat) {
-        throw new Error('Merrkat not connected.');
-      }
-
-      const expApiTypeMapping = createTypeMapping(
-        this._cip30ExperimentalApi ?? new ExperimentalContainer<any>({})
-      );
-      const expFullApiTypeMapping = createTypeMapping(
-        this._cip30EnableExperimentalApi ?? new ExperimentalContainer<any>({})
-      );
-
-      let args = {
-        api: {
-          apiVersion: this.walletInfo.version,
-          name: this.walletInfo.name,
-          icon: this.walletInfo.icon,
-          methods: cip30Functions,
-          experimentalApi: serializeTypeMapping(expApiTypeMapping),
-          fullExperimentalApi: serializeTypeMapping(expFullApiTypeMapping),
-        },
-      };
-
-      this.meerkat.rpc(
-        identifier,
-        'api',
-        args,
-        (connectMessage: IConnectMessage) => {
-          if (!this.meerkat) {
-            throw new Error('Meerkat not connected.');
-          }
-
-          if (connectMessage.error) {
-            this.meerkat.logger.warn(
-              'Api could note be injected. Error: ' +
-                connectMessage.errorMessage
-                ? connectMessage.errorMessage
-                : 'unknown error.'
-            );
-          }
-
-          this.onApiInject(connectMessage);
-        }
-      );
-    };
-
-    // https://cips.cardano.org/cips/cip30/
-    const cip30Functions: Array<Cip30Function> = [
-      'getNetworkId',
-      'getUtxos',
-      'getCollateral',
-      'getBalance',
-      'getUsedAddresses',
-      'getUnusedAddresses',
-      'getChangeAddress',
-      'getRewardAddresses',
-      'signTx',
-      'signData',
-      'submitTx',
-    ];
-
-    this.meerkat.on('server', () => {
-
-      this.meerkat?.logger.debug('WALLET: DApp server seen, create connection!');
-
-      if(!this.meerkat) {
-
-        throw new Error('Meerkat not connected.');
-      }
-
-      this.meerkat.rpc(identifier, 'connect', this.walletInfo, (connectStatus: IConnectMessage) => {
-
-        this.meerkat?.logger.debug('WALLET: DApp now connected with status:', connectStatus);
-
-        if (connectStatus.connected) {
-
-          injectApi();
-
-          if(this.DAppDiscoveryMeerkat) {
-
-            //close discovery meerkat as we are connected
-            this.meerkat?.logger.debug('WALLET: Close discovery after successfully connected to server.');
-            this.DAppDiscoveryMeerkat.close();
-
-          }
-        } else {
-
-          if(!this.meerkat) {
-
-            throw new Error('Merrkat not connected.');
-          }
-
-            this.meerkat.logger.warn(
-              'Connection failed. Another wallet has already been connected to this dApp.'
-            );
-          }
-
-          this.generateIdenticon();
-
-        if(this.DAppDiscoveryMeerkat?.address()) {
-
-          this.meerkat!.rpc(identifier, 'setDiscovery', { walletDiscoveryAddress: this.DAppDiscoveryMeerkat.address() }, (connectStatus: boolean) => {
-
-            this.meerkat?.logger.debug('WALLET: result of setDiscovery call', connectStatus)
-          })
-
-        } else {
-
-          this.meerkat?.logger.debug('WALLET: discovery meerkat has no address ?')
-        }
-
-        this.meerkat?.logger.debug('WALLET: calling onConnect event callback from wallet!', this.onConnect)
-
-        this.onConnect(connectStatus)
-      });
-    });
-
-    this.meerkats.push(this.meerkat);
-    return this.meerkat.seed;
-  }
-
-  public generateIdenticon = () => {
-    if (!this.meerkat?.address()) {
-      throw new Error('Server meerkat address not defined.');
-    }
-
-    if (!this.meerkat?.identifier) {
-      throw new Error('Client meerkat address not defined.');
-    }
-
-    this.identicon = PeerConnectIdenticon.getBase64Identicon(
-      this.meerkat?.address() + this.meerkat?.identifier
-    );
-  };
-
-  public disconnect(address: string) {
-    if (!this.meerkat) {
-      throw new Error('Meerkat not connected.');
-    }
-
-    this.meerkat.rpc(
-      address,
-      'disconnect',
-      this.walletInfo,
-      (connectStatus: IConnectMessage) => {
-        if (this.meerkat) {
-          this.meerkat.close();
-        }
-
-      this.clearSeen();
-      this.onDisconnect(connectStatus);
-
-      this.meerkat?.logger.debug('WALLET: setup discovery after disconnect was called');
-      this.setUpDiscoveryMeerkat();
-    })
-  }
-
-  public getIdenticon = () => {
-    return this.identicon;
-  };
-
-  protected abstract getNetworkId(): Promise<number>;
-  protected abstract getUtxos(
-    amount?: Cbor,
-    paginate?: Paginate
-  ): Promise<Cbor[] | null>;
-  protected abstract getCollateral(params?: {
-    amount?: Cbor;
-  }): Promise<Cbor[] | null>;
-  protected abstract getBalance(): Promise<Cbor>;
-  protected abstract getUsedAddresses(): Promise<Cbor[]>;
-  protected abstract getUnusedAddresses(): Promise<Cbor[]>;
-  protected abstract getChangeAddress(): Promise<Cbor>;
-  protected abstract getRewardAddresses(): Promise<Cbor[]>;
-  protected abstract signTx(tx: Cbor, partialSign: boolean): Promise<Cbor>;
-  protected abstract signData(
-    addr: string,
-    payload: Bytes
-  ): Promise<Cip30DataSignature>;
-  protected abstract submitTx(tx: Cbor): Promise<string>;
-}
-
-class PeerConnectIdenticon {
-  public static getBase64Identicon = (hash: string): string | null => {
-    if (hash.length < 68) {
-      console.warn(
-        'Meerkat connection hash is to short. Not generating identicon.'
-      );
-
-      return null;
-    }
-
-    return identicon(
-      hash
-        .split('')
-        .reverse()
-        .map((char: string, index: number) =>
-          index > 0 && index % 10 === 0 ? '-' : char
-        )
-        .join(''),
-      {
-        size: 100,
-        baseSeed: 'cardano-peer-connect',
-        fontSize: 0.17,
-        initialsColours: ['#000000', '#FF0000', '#0000FF'],
-      }
-    ).toDataURL();
-  };
-}
-
-class AutoConnectHelper {
-
-  private static storageKey = 'cardano-peer-autoconnect-id';
-  private static discoveryStorageKey = 'cardano-peer-discovery-id';
-  private static walletDiscoveryStorageKey = 'cardano-wallet-discovery-address';
-
-  public static addAutoConnectId = (id: string): void => {
-    let autoConnectIds = [];
-
-    const ids = localStorage.getItem(this.storageKey);
-
-    if (ids !== null) {
-      autoConnectIds = JSON.parse(ids);
-    }
-
-    if (this.isAutoConnectId(id)) {
-      return;
-    }
-
-    autoConnectIds.push(id);
-
-    localStorage.setItem(this.storageKey, JSON.stringify(autoConnectIds));
-  };
-
-  public static getAutoConnectIds = (): string[] => {
-    return JSON.parse(localStorage.getItem(this.storageKey) ?? '[]');
-  };
-
-  public static isAutoConnectId = (id: string): boolean => {
-    return this.getAutoConnectIds().includes(id);
-  };
-
-  public static resetAutoConnectIds = (): void => {
-    localStorage.setItem(this.storageKey, JSON.stringify([]));
-  };
-
-  public static removeAutoConnectId = (id: string): void => {
-    let autoConnectIds = [];
-    const ids = localStorage.getItem(this.storageKey);
-
-    if (ids !== null) {
-      autoConnectIds = JSON.parse(ids);
-    }
-
-    const index = autoConnectIds.indexOf(id);
-
-    if (index !== -1) {
-      autoConnectIds = autoConnectIds.splice(index, 1);
-
-      localStorage.setItem(this.storageKey, JSON.stringify(autoConnectIds));
-      return;
-    }
-  }
-
-  public static saveWalletAutoDiscoverySeed = (id: string):void => {
-
-    localStorage.setItem(this.discoveryStorageKey, id);
-  }
-
-  public static getWalletAutoDiscoverySeed = (): string | null => {
-
-    return localStorage.getItem(this.discoveryStorageKey)
-  }
-
-
-  public static saveWalletDiscoveryAddress = (id: string):void => {
-
-    localStorage.setItem(this.walletDiscoveryStorageKey, id);
-  }
-
-  public static getWalletDiscoveryAddress = (): string | null => {
-
-    return localStorage.getItem(this.walletDiscoveryStorageKey)
-  }
-
-}
-
-/**
- * ToDo: replace with full Logger implementation.
- * The logger in meerkat could not be used, therefore this is a minimal logger implementation.
- */
-class InfoLogger {
-
-  protected level = 'info'
-  protected enable = true
-  protected name = "InfoLogger"
-
-  public constructor(name: string, enable: boolean = true, level: string = 'info') {
-    this.name = name
-    this.level = level
-    this.enable = enable
-  }
-
-  public setLevel = (level: string) => {
-
-    this.level = level
-  }
-
-  public info = (...args : Array<any>) => {
-
-    this.doLog('info', ...args)
-  }
-
-  public debug = (...args : Array<any>) => {
-
-    if(this.level === 'debug') {
-      this.doLog('debug', ...args)
-    }
-  }
-
-  public warn = (...args : Array<any>) => {
-
-    this.doLog('warn', ...args)
-  }
-
-  public error = (...args : Array<any>) => {
-
-    this.doLog('error', ...args)
-  }
-
-  public doLog = (level: string, ...message : Array<any> ) => {
-    if(this.enable) {
-      console.log(`${this.level}:`, ...message)
-    }
-  }
-
-}
-
-export {
-  DAppPeerConnect,
-  CardanoPeerConnect,
-  PeerConnectIdenticon,
-  AutoConnectHelper,
-  ExperimentalContainer,
-};
+export { default as AutoConnectHelper } from './src/lib/AutoConnectHelper';
+export { default as PeerConnectIdenticon } from './src/lib/PeerConnectIdenticon';
+export { ExperimentalContainer } from './src/lib/ExperimentalContainer';
+export { default as CardanoPeerConnect } from './src/CardanoPeerConnect';
+export { default as DAppPeerConnect } from './src/DAppPeerConnect';
